Remove deprecated un-used event notifiers
These have been replaced by the 3 new event listeners for:

visitor-linked / visitor-request-approved / visitor-request-rejected
