--- src/test/kotlin/uk/gov/justice/digital/hmpps/notificationsalertsvsip/integration/domainevents/email/BookerVisitorApprovedEventEmailTest.kt
@@ -1,191 +1,0 @@
-package uk.gov.justice.digital.hmpps.notificationsalertsvsip.integration.domainevents.email
-
-import org.awaitility.kotlin.await
-import org.awaitility.kotlin.untilAsserted
-import org.junit.jupiter.api.Test
-import org.mockito.Mockito
-import org.mockito.kotlin.any
-import org.mockito.kotlin.times
-import org.mockito.kotlin.verify
-import org.springframework.http.HttpStatus
-import uk.gov.justice.digital.hmpps.notificationsalertsvsip.dto.booker.registry.BookerInfoDto
-import uk.gov.justice.digital.hmpps.notificationsalertsvsip.dto.booker.registry.VisitorRequestVisitorInfoDto
-import uk.gov.justice.digital.hmpps.notificationsalertsvsip.dto.prisoner.contact.registry.ContactWithOptionalPrisonerRelationshipDto
-import uk.gov.justice.digital.hmpps.notificationsalertsvsip.enums.EmailTemplateNames
-import uk.gov.justice.digital.hmpps.notificationsalertsvsip.enums.LanguagePreference
-import uk.gov.justice.digital.hmpps.notificationsalertsvsip.enums.booker.registry.BookerEventType
-import uk.gov.justice.digital.hmpps.notificationsalertsvsip.integration.domainevents.EventsIntegrationTestBase
-import uk.gov.justice.digital.hmpps.notificationsalertsvsip.service.listeners.events.additionalinfo.VisitorLinkedAdditionalInfo
-import uk.gov.justice.digital.hmpps.notificationsalertsvsip.service.listeners.notifiers.BOOKER_VISITOR_APPROVED
-import java.time.LocalDate
-
-class BookerVisitorApprovedEventEmailTest : EventsIntegrationTestBase() {
-  private val prisonerId = "A1234BC"
-  private val bookerReference = "booker-ref"
-  private val bookerEmailAddress = "test@example.com"
-  private val contact1 = ContactWithOptionalPrisonerRelationshipDto(1234, "Visitor", "One", (LocalDate.now().minusYears(30)))
-
-  @Test
-  fun `when visitor approved event is received a visitor approved email is sent to the booker`() {
-    // Given
-    val visitorId = 1234L
-    val templateId = notificationTemplateResolver.getEmailTemplate(EmailTemplateNames.BOOKER_VISITOR_APPROVED, LanguagePreference.EN)
-    val templateVars = mapOf(
-      "visitor" to contact1.firstName + " " + contact1.lastName,
-    )
-
-    val bookerInfo = BookerInfoDto(bookerReference, bookerEmailAddress)
-    val bookerAdditionalInfo = VisitorLinkedAdditionalInfo(bookerReference, prisonerId, visitorId.toString())
-    val domainEvent = createDomainEventJson(BOOKER_VISITOR_APPROVED, createAdditionalInformationJson(bookerAdditionalInfo))
-    val jsonSqsMessage = createSQSMessage(domainEvent)
-
-    Mockito.`when`(
-      notificationClient.sendEmail(
-        templateId,
-        bookerInfo.email,
-        templateVars,
-        null,
-        "00000000-0000-0000-0000-000000000001",
-      ),
-    ).thenReturn(buildSendEmailResponse(reference = "test"))
-
-    // When
-    domainEventListenerService.onDomainEvent(jsonSqsMessage)
-    bookerRegistryMockServer.stubGetBooker(bookerReference, bookerInfo)
-    prisonerContactRegisterMockServer.stubSearchContacts(prisonerId, listOf(visitorId), false, listOf(contact1))
-
-    // Then
-    verifyBookerEmailSent(templateId, bookerAdditionalInfo, bookerInfo, VisitorRequestVisitorInfoDto(contact1), templateVars)
-  }
-
-  @Test
-  fun `when visitor approved event is received but visitor is not on the social contacts an email is not sent to the booker`() {
-    // Given
-    // visitor does not exist on the prisoner's social contacts list
-    val visitorId = 9999L
-
-    val bookerInfo = BookerInfoDto(bookerReference, bookerEmailAddress)
-    val bookerAdditionalInfo = VisitorLinkedAdditionalInfo(bookerReference, prisonerId, visitorId.toString())
-    val domainEvent = createDomainEventJson(BOOKER_VISITOR_APPROVED, createAdditionalInformationJson(bookerAdditionalInfo))
-    val jsonSqsMessage = createSQSMessage(domainEvent)
-
-    // When
-    domainEventListenerService.onDomainEvent(jsonSqsMessage)
-    bookerRegistryMockServer.stubGetBooker(bookerReference, bookerInfo)
-    prisonerContactRegisterMockServer.stubSearchContacts(prisonerId, listOf(visitorId), false, emptyList())
-
-    // Then
-    verifyBookerEmailNotSent(bookerAdditionalInfo)
-  }
-
-  @Test
-  fun `when visitor approved event is received but prisoner contact search returns an INTERNAL_SERVER error an email is not sent to the booker`() {
-    // Given
-    // visitor does not exist on the prisoner's social contacts list
-    val visitorId = 9999L
-
-    val bookerInfo = BookerInfoDto(bookerReference, bookerEmailAddress)
-    val bookerAdditionalInfo = VisitorLinkedAdditionalInfo(bookerReference, prisonerId, visitorId.toString())
-    val domainEvent = createDomainEventJson(BOOKER_VISITOR_APPROVED, createAdditionalInformationJson(bookerAdditionalInfo))
-    val jsonSqsMessage = createSQSMessage(domainEvent)
-
-    // When
-    domainEventListenerService.onDomainEvent(jsonSqsMessage)
-    bookerRegistryMockServer.stubGetBooker(bookerReference, bookerInfo)
-    prisonerContactRegisterMockServer.stubSearchContacts(prisonerId, listOf(visitorId), false, null, HttpStatus.INTERNAL_SERVER_ERROR)
-
-    // Then
-    verifyBookerEmailNotSent(bookerAdditionalInfo)
-  }
-
-  @Test
-  fun `when visitor approved event is received but booker registry returns a NOT_FOUND error an email is not sent to the booker`() {
-    // Given
-    val visitorId = 1234L
-
-    val bookerAdditionalInfo = VisitorLinkedAdditionalInfo(bookerReference, prisonerId, visitorId.toString())
-    val domainEvent = createDomainEventJson(BOOKER_VISITOR_APPROVED, createAdditionalInformationJson(bookerAdditionalInfo))
-    val jsonSqsMessage = createSQSMessage(domainEvent)
-
-    // When
-    domainEventListenerService.onDomainEvent(jsonSqsMessage)
-    bookerRegistryMockServer.stubGetBooker(bookerReference, null, HttpStatus.NOT_FOUND)
-    prisonerContactRegisterMockServer.stubSearchContacts(prisonerId, listOf(visitorId), false, listOf(contact1))
-
-    // Then
-    verifyBookerEmailNotSent(bookerAdditionalInfo)
-  }
-
-  @Test
-  fun `when visitor approved event is received but booker registry returns an INTERNAL_SERVER error an email is not sent to the booker`() {
-    // Given
-    val visitorId = 1234L
-
-    val bookerAdditionalInfo = VisitorLinkedAdditionalInfo(bookerReference, prisonerId, visitorId.toString())
-    val domainEvent = createDomainEventJson(BOOKER_VISITOR_APPROVED, createAdditionalInformationJson(bookerAdditionalInfo))
-    val jsonSqsMessage = createSQSMessage(domainEvent)
-
-    // When
-    domainEventListenerService.onDomainEvent(jsonSqsMessage)
-    bookerRegistryMockServer.stubGetBooker(bookerReference, null, HttpStatus.INTERNAL_SERVER_ERROR)
-    prisonerContactRegisterMockServer.stubSearchContacts(prisonerId, listOf(visitorId), false, listOf(contact1))
-
-    // Then
-    verifyBookerEmailNotSent(bookerAdditionalInfo)
-  }
-
-  @Test
-  fun `when visitor approved message is received and language is welsh but no welsh template exists, a visitor approved email is sent in english`() {
-    // Given
-    val visitorId = 1234L
-    val templateId = notificationTemplateResolver.getEmailTemplate(EmailTemplateNames.BOOKER_VISITOR_APPROVED, LanguagePreference.CY)
-    val templateVars = mapOf(
-      "visitor" to contact1.firstName + " " + contact1.lastName,
-    )
-
-    val bookerInfo = BookerInfoDto(bookerReference, bookerEmailAddress)
-    val bookerAdditionalInfo = VisitorLinkedAdditionalInfo(bookerReference, prisonerId, visitorId.toString())
-    val domainEvent = createDomainEventJson(BOOKER_VISITOR_APPROVED, createAdditionalInformationJson(bookerAdditionalInfo))
-    val jsonSqsMessage = createSQSMessage(domainEvent)
-
-    Mockito.`when`(
-      notificationClient.sendEmail(
-        templateId,
-        bookerInfo.email,
-        templateVars,
-        null,
-        "00000000-0000-0000-0000-000000000001",
-      ),
-    ).thenReturn(buildSendEmailResponse(reference = "test"))
-
-    // When
-    domainEventListenerService.onDomainEvent(jsonSqsMessage)
-    bookerRegistryMockServer.stubGetBooker(bookerReference, bookerInfo)
-    prisonerContactRegisterMockServer.stubSearchContacts(prisonerId, listOf(visitorId), false, listOf(contact1))
-
-    // Then
-    verifyBookerEmailSent(templateId, bookerAdditionalInfo, bookerInfo, VisitorRequestVisitorInfoDto(contact1), templateVars)
-  }
-
-  private fun verifyBookerEmailSent(templateId: String, additionalInfo: VisitorLinkedAdditionalInfo, bookerInfoDto: BookerInfoDto, visitorInfo: VisitorRequestVisitorInfoDto, templateVars: Map<String, Any>) {
-    await untilAsserted { verify(bookerVisitorApprovedEventNotifierSpy, times(1)).processEvent(any()) }
-    await untilAsserted { verify(visitorRequestNotificationService, times(1)).sendVisitorLinkedEmail(additionalInfo) }
-    await untilAsserted { verify(emailSenderService, times(1)).sendBookerVisitorEmail(bookerInfoDto, visitorInfo, BookerEventType.VISITOR_APPROVED, "00000000-0000-0000-0000-000000000001") }
-
-    await untilAsserted {
-      verify(notificationClient, times(1)).sendEmail(
-        templateId,
-        bookerInfoDto.email,
-        templateVars,
-        null,
-        "00000000-0000-0000-0000-000000000001",
-      )
-    }
-  }
-
-  private fun verifyBookerEmailNotSent(additionalInfo: VisitorLinkedAdditionalInfo) {
-    await untilAsserted { verify(bookerVisitorApprovedEventNotifierSpy, times(1)).processEvent(any()) }
-    await untilAsserted { verify(visitorRequestNotificationService, times(1)).sendVisitorLinkedEmail(additionalInfo) }
-    await untilAsserted { verify(emailSenderService, times(0)).sendBookerVisitorEmail(any(), any(), any(), any()) }
-  }
-}
