+++ src/test/kotlin/uk/gov/justice/digital/hmpps/notificationsalertsvsip/integration/domainevents/email/BookerVisitorRequestRejectedEventEmailTest.kt
@@ -79,6 +79,65 @@ class BookerVisitorRequestRejectedEventEmailTest : EventsIntegrationTestBase() {
     }
   }
 
+  @Test
+  fun `when visitor request rejected event is received with reason already linked a visitor rejected already linked email is sent to the booker`() {
+    // Given
+    val prisonerId = "A1234BC"
+    val bookerReference = "booker-ref"
+    val bookerEmailAddress = "test@example.com"
+    val visitorRequestReference = "abc-def-ghi"
+    val visitorRequest = createVisitorRequest(
+      visitorRequestReference = visitorRequestReference,
+      bookerReference = bookerReference,
+      bookerEmailAddress = bookerEmailAddress,
+      prisonerId = prisonerId,
+      rejectionReason = "ALREADY_LINKED",
+    )
+    val templateId = notificationTemplateResolver.getEmailTemplate(EmailTemplateNames.BOOKER_VISITOR_REJECTED_ALREADY_LINKED, LanguagePreference.EN)
+    val templateVars = mapOf(
+      "visitor" to "${visitorRequest.firstName} ${visitorRequest.lastName}",
+    )
+    val bookerInfo = BookerInfoDto(bookerReference, bookerEmailAddress)
+    val bookerAdditionalInfo = VisitorRequestAdditionalInfo(visitorRequestReference)
+    val domainEvent = createDomainEventJson(BOOKER_VISITOR_REQUEST_REJECTED, createAdditionalInformationJson(bookerAdditionalInfo))
+    val jsonSqsMessage = createSQSMessage(domainEvent)
+
+    Mockito.`when`(
+      notificationClient.sendEmail(
+        templateId,
+        bookerInfo.email,
+        templateVars,
+        null,
+        "00000000-0000-0000-0000-000000000001",
+      ),
+    ).thenReturn(buildSendEmailResponse(reference = "test"))
+
+    // When
+    bookerRegistryMockServer.stubGetVisitorRequestByReference(visitorRequestReference, visitorRequest)
+    domainEventListenerService.onDomainEvent(jsonSqsMessage)
+
+    // Then
+    await untilAsserted { verify(bookerVisitorRequestRejectedEventNotifierSpy, times(1)).processEvent(any()) }
+    await untilAsserted { verify(visitorRequestNotificationService, times(1)).sendVisitorRequestRejectedEmail(bookerAdditionalInfo) }
+    await untilAsserted {
+      verify(emailSenderService, times(1)).sendBookerVisitorEmail(
+        bookerInfo,
+        VisitorRequestVisitorInfoDto(visitorRequest),
+        BookerEventType.VISITOR_REJECTED_ALREADY_LINKED,
+        "00000000-0000-0000-0000-000000000001",
+      )
+    }
+    await untilAsserted {
+      verify(notificationClient, times(1)).sendEmail(
+        templateId,
+        bookerInfo.email,
+        templateVars,
+        null,
+        "00000000-0000-0000-0000-000000000001",
+      )
+    }
+  }
+
   private fun createVisitorRequest(
     visitorRequestReference: String,
     bookerReference: String,
